requirements.txt
# pip install -r requirements.txt

# base ----------------------------------------
Cython
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.2
Pillow
PyYAML>=5.3.1
scipy>=1.4.1
tensorboard>=2.2
torch>=1.9.0,<1.12.0
torchvision>=0.10.0,<0.13.0
tqdm>=4.41.0

# logging -------------------------------------
# wandb

# plotting ------------------------------------
seaborn>=0.11.0
pandas

# export --------------------------------------
# coremltools==4.0
onnx>=1.8.0
#scikit-learn==0.19.2  # for coreml quantization
yacs
tabulate
Ninja
#for TensorRT testing
pycuda


# extras --------------------------------------
thop  # FLOPS computation
pycocotools>=2.0  # COCO mAP